% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gknown.R
\name{bayes_men_g4}
\alias{bayes_men_g4}
\title{Bayesian test for segregation distortion in tetraploids when genotypes are known.}
\usage{
bayes_men_g4(
  x,
  g1,
  g2,
  drbound = 1/6,
  shape1 = 5/9,
  shape2 = 10/9,
  ts1 = 1,
  ts2 = 1,
  pp = TRUE,
  dr = TRUE,
  xi1 = 1/3,
  xi2 = 1/3,
  alpha = 0,
  ...
)
}
\arguments{
\item{x}{A vector of genotype counts. \code{x[i]} is the number of
offspring with genotype \code{i-1}.}

\item{g1}{The genotype of parent 1.}

\item{g2}{The genotype of parent 2.}

\item{drbound}{The maximum rate of double reduction. A default of 1/6
is provided, which is the rate under the complete equational
segregation model of meiosis.}

\item{shape1}{The shape 1 parameter for the beta prior over the
preferential pairing parameter.}

\item{shape2}{The shape 2 parameter for the beta prior over the
preferential pairing parameter.}

\item{ts1}{The shape 1 parameter for the beta prior over the
rate of quadrivalent formation.}

\item{ts2}{The shape 2 parameter for the beta prior over the
rate of quadrivalent formation.}

\item{pp}{A logical. Should we account for preferential pairing
(\code{TRUE}) or not (\code{FALSE})?}

\item{dr}{A logical. Should we account for double reduction
(\code{TRUE}) or not (\code{FALSE})?}

\item{xi1}{If \code{pp = FALSE}, then \code{xi} is the known rate of
preferential pairing (assumed fixed) of parent 1.}

\item{xi2}{If \code{pp = FALSE}, then \code{xi} is the known rate of
preferential pairing (assumed fixed) of parent 2.}

\item{alpha}{If \code{dr = FALSE}, then \code{alpha} is the known rate of
double reduction (assumed fixed).}

\item{...}{Additional arguments to pass to Stan.}
}
\value{
A list with the following elements
\describe{
\item{\code{lbf}}{The log Bayes Factor.}
\item{\code{m0}}{Marginal likelihood under null.}
\item{\code{ma}}{Marginal likelihood under alternative.}
\item{\code{alpha}}{The posterior mean double reduction rate.}
\item{\code{xi1}}{The posterior mean preferential pairing parameter of parent 1.}
\item{\code{xi2}}{The posterior mean preferential pairing parameter of parent 2.}
}
}
\description{
This will run a Bayesian test using the genotypes of an F1 population
of tetraploids for the null of Mendelian segregation (accounting for double
reduction and preferential pairing) against the alternative of
segregation distortion. This is when the genotypes are assumed known.
}
\section{Impossible genotypes}{

Some offspring genotype combinations are impossible given the parental
gentoypes. If these impossible genotypes combinations show up, we return a
log-BF of \code{-Inf} and missing values for all other return items.
The impossible genotypes are (when double reduction is allowed):
\describe{
\item{\code{g1 == 0 && g2 == 0}}{Only offspring genotypes of 0 are possible.}
\item{\code{g1 == 4 && g2 == 4}}{Only offspring genotypes of 4 are possible.}
\item{\code{g1 == 0 && g2 == 4 || g1 == 4 && g2 == 0}}{Only offspring genotypes of 2 are possible.}
\item{\code{g1 == 0 && g2 \%in\% c(1, 2, 3) || g1 \%in\% c(1, 2, 3) && g2 == 0}}{Only offspring genotypes of 0, 1, and 2 are possible.}
\item{\code{g1 == 4 && g2 \%in\% c(1, 2, 3) || g1 \%in\% c(1, 2, 3) && g2 == 4}}{Only offspring genotypes of 2, 3, and 4 are possible.}
}
When double reduction is 0, the list of impossible genotypes is more
complicated.
}

\section{Unidentified parameters}{

When \code{g1 = 2} or \code{g2 = 2} (or both), the model is not identified
and those estimates (\code{alpha}, \code{xi1}, and \code{xi2}) are
meaningless. Do NOT interpret them.

The estimate of \code{alpha} (double reduction rate) IS identified as
long as at least one parent is simplex, and no parent is duplex.
However, the estimates of the double reduction rate have extremely high
variance.
}

\examples{
set.seed(1)
gf <- offspring_gf_2(alpha = 1/12, xi1 = 0.1, xi2 = 0.5, p1 = 2, p2 = 2)
x <- offspring_geno(gf = gf, n = 20)
bayes_men_g4(x = x, g1 = 2, g2 = 2, chains = 1, iter = 1000)

}
\author{
David Gerard
}
