% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gknown.R
\name{chisq_ndr_npp_g4}
\alias{chisq_ndr_npp_g4}
\title{Chi Square test when genotypes are known}
\usage{
chisq_ndr_npp_g4(y, l1, l2)
}
\arguments{
\item{y}{Vector of observed genotype counts}

\item{l1}{Parent 1's genotype}

\item{l2}{Parent 2's genotype}
}
\value{
The Chi Square statistic and p-value
}
\description{
This chi-squared test is run under the assumption of no double reduction
and no preferential pairing.
}
\examples{
y <- c(1, 2, 4, 3, 0)
l1 <- 2
l2 <- 2
chisq_ndr_npp_g4(y, l1, l2)

y <- c(10, 25, 10, 0, 0)
l1 <- 1
l2 <- 1
chisq_ndr_npp_g4(y, l1, l2)

}
\author{
Mira Thakkar and David Gerard
}
