% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_gknown.R
\name{find_df}
\alias{find_df}
\title{Calculate degrees of freedom of test}
\usage{
find_df(alpha, xi1, xi2, g1, g2, drbound = 1/6)
}
\arguments{
\item{alpha}{estimated double reduction rate}

\item{xi1}{estimated preferential pairing parameter of parent 1}

\item{xi2}{estimated preferential pairing parameter of parent 2}

\item{g1}{genotype of parent 1.}

\item{g2}{genotype of parent 2}

\item{drbound}{the maximum possible value of alpha}
}
\description{
Calculate degrees of freedom of test
}
\author{
David Gerard
}
