% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_glpknown.R
\name{like_glpknown}
\alias{like_glpknown}
\title{Likelihood under three parameter model when using offspring genotypes
likelihoods but parent genotypes are known.}
\usage{
like_glpknown(gl, tau, beta, gamma1, gamma2, g1, g2, log_p = TRUE)
}
\arguments{
\item{gl}{The matrix of genotype likelihoods of the offspring. Rows index
The individuals, columns index the genotypes.}

\item{beta}{The probability of double reduction given quadrivalent formation.}

\item{gamma1}{The probability of AA_aa pairing for parent 1.}

\item{gamma2}{The probability of AA_aa pairing for parent 2.}

\item{g1}{Parent 1's genotype.}

\item{g2}{Parent 2's genotype.}

\item{log_p}{A logical. Should we return the log likelihood or not?}
}
\description{
Likelihood under three parameter model when using offspring genotypes
likelihoods but parent genotypes are known.
}
\examples{
p1 <- 1
p2 <- 0
gf <- offspring_gf_2(alpha = 0, xi1 = 1/3, xi2 = 1/3, p1 = p1, p2 = p2)
x <- offspring_geno(gf = gf, n = 10)
genovec <- gcount_to_gvec(gcount = x)
uout <- po_gl(genovec = genovec, p1_geno = p1, p2_geno = p2, ploidy = 4)
gl <- uout$genologlike
like_glpknown(gl = gl, tau = 1/2, beta = 1/12, gamma1 = 1/3, gamma2 = 1/3, g1 = p1, g2 = p2, log_p = TRUE)


}
\author{
David Gerard
}
