% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_gknown.R
\name{lrt_dr_pp_g4}
\alias{lrt_dr_pp_g4}
\title{LRT when both double reduction and preferential pairing are not known.}
\usage{
lrt_dr_pp_g4(x, g1, g2, drbound = 1/6)
}
\arguments{
\item{x}{A vector of length 5. \code{x[i]} is the count of individuals with
genotype \code{i-1}.}

\item{g1}{Parent 1's genotype.}

\item{g2}{Parent 2's genotype.}

\item{drbound}{the upper bound on the double reduction rate.}
}
\description{
All genotypes are known.
}
\section{Impossible genotypes}{


Some offspring genotype combinations are impossible given the parental
gentoypes. If these impossible genotypes combinations show up, we return a
p-value of 0, a log-likelihood ratio statistic of Infinity, and missing
values for all other return items. The impossible genotypes are:
\describe{
\item{\code{g1 = 0 && g2 = 0}}{Only offspring genotypes of 0 are possible.}
\item{\code{g1 = 4 && g2 = 4}}{Only offspring genotypes of 4 are possible.}
\item{\code{g1 = 0 && g2 = 4 || g1 == 4 && g2 == 0}}{Only offspring genotypes of 2 are possible.}
\item{\code{g1 = 0 && g2 \%in\% c(1, 2, 3) || g1 = \%in\% c(1, 2, 3) && g2 == 0}}{Only offspring genotypes of 0, 1, and 2 are possible.}
\item{\code{g1 = 4 && g2 %in% c(1, 2, 3) || g1 = %in% c(1, 2, 3) && g2 == 4}}{Only offspring genotypes of 2, 3, and 4 are possible.}
}
}

\examples{
alpha <- 1/12
xi1 <- 1/3
xi2 <- 1/3
n <- 1000

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 0, p2 = 0)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 0, g2 = 0)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 1, p2 = 0)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 1, g2 = 0)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 1, p2 = 1)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 1, g2 = 1)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 1, p2 = 2)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 1, g2 = 2)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 1, p2 = 3)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 1, g2 = 3)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 1, p2 = 4)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 1, g2 = 4)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 2, p2 = 0)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 2, g2 = 0)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 2, p2 = 1)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 2, g2 = 1)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 2, p2 = 2)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 2, g2 = 2)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 2, p2 = 3)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 2, g2 = 3)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 2, p2 = 4)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 2, g2 = 4)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 3, p2 = 0)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 3, g2 = 0)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 3, p2 = 1)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 3, g2 = 1)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 3, p2 = 2)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 3, g2 = 2)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 3, p2 = 3)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 3, g2 = 3)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 3, p2 = 4)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 3, g2 = 4)

gf <- offspring_gf_2(alpha = alpha, xi1 = xi1, xi2 = xi2, p1 = 4, p2 = 4)
x <- offspring_geno(gf = gf, n = n)
lrt_dr_pp_g4(x = x, g1 = 4, g2 = 4)
}
\author{
David Gerard
}
