% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_glpknown.R
\name{lrt_dr_pp_glpknown4}
\alias{lrt_dr_pp_glpknown4}
\title{LRT when both double reduction and preferential pairing are not known.}
\usage{
lrt_dr_pp_glpknown4(gl, g1, g2, drbound = 1/6)
}
\arguments{
\item{gl}{The matrix of genotype likelihoods of the offspring. Rows index
The individuals, columns index the genotypes.}

\item{g1}{Parent 1's genotype.}

\item{g2}{Parent 2's genotype.}

\item{drbound}{Maximum value of double reduction.}
}
\description{
Offspring use genotype likelihoods, parent genotypes are known.
}
\author{
David Gerard
}
