% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_glpknown.R
\name{lrt_men_gl4}
\alias{lrt_men_gl4}
\title{Likelihood ratio test using genotype likelihoods.}
\usage{
lrt_men_gl4(gl, g1, g2, drbound = 1/6, pp = TRUE, dr = TRUE)
}
\arguments{
\item{gl}{The genotype log-likelihoods. The rows index the individuals
and the columns index the genotypes.}

\item{g1}{Either parent 1's genotype, or parent 1's genotype log-likelihoods.}

\item{g2}{Either parent 2's genotype, or parent 2's genotype log-likelihoods.}

\item{drbound}{The upper bound on the double reduction rate.}

\item{pp}{Is (partial) preferential pairing possible (\code{TRUE}) or
not (\code{FALSE})?}

\item{dr}{Is double reduction possible (\code{TRUE}) or
not (\code{FALSE})?}
}
\description{
Likelihood ratio test using genotype likelihoods.
}
\examples{
## null sim
set.seed(10)
g1 <- 2
g2 <- 2
gl <- simf1gl(n = 25, g1 = g1, g2 = g2, alpha = 1/12, xi2 = 1/4)
lrt_men_gl4(gl = gl, g1 = g1, g2 = g2)

## Alt sim
gl <- hwep::simgl(nvec = rep(5, 5))
lrt_men_gl4(gl = gl, g1 = g1, g2 = g2)


}
\author{
David Gerard
}
