% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_gknown.R
\name{lrt_ndr_npp_g4}
\alias{lrt_ndr_npp_g4}
\title{Likelihood ratio test assuming no double reduction and no preferential pairing}
\usage{
lrt_ndr_npp_g4(x, g1, g2)
}
\arguments{
\item{x}{A vector of length 5. \code{x[i]} is the count of individuals with
genotype \code{i-1}.}

\item{g1}{Parent 1's genotype.}

\item{g2}{Parent 2's genotype.}
}
\value{
A list of length three with the following elements
\describe{
\item{\code{statistic}}{The likelihood ratio test statistic.}
\item{\code{df}}{The degrees of freedom.}
\item{\code{p_value}}{The p-value.}
}
}
\description{
Likelihood ratio test assuming no double reduction and no preferential pairing
}
\author{
David Gerard
}
