% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_glpknown.R
\name{lrt_ndr_npp_glpknown4}
\alias{lrt_ndr_npp_glpknown4}
\title{Likelihood ratio test assuming no double reduction and no preferential pairing}
\usage{
lrt_ndr_npp_glpknown4(gl, g1, g2)
}
\arguments{
\item{gl}{The matrix of genotype likelihoods of the offspring. Rows index
The individuals, columns index the genotypes.}

\item{g1}{Parent 1's genotype.}

\item{g2}{Parent 2's genotype.}
}
\value{
A list of length three with the following elements
\describe{
\item{\code{statistic}}{The likelihood ratio test statistic.}
\item{\code{df}}{The degrees of freedom.}
\item{\code{p_value}}{The p-value.}
}
}
\description{
This is when offspring genotypes are not known
}
\author{
David Gerard
}
