% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.R
\name{marg_f1_dr_npp_gl4}
\alias{marg_f1_dr_npp_gl4}
\title{Marginal likelihood, double reduction, no preferential pairing, genotype likelihoods.}
\usage{
marg_f1_dr_npp_gl4(
  gl,
  p1_gl = rep(log(0.2), 5),
  p2_gl = rep(log(0.2), 5),
  mixprop = 0.001,
  lg = TRUE,
  output = c("marg", "all"),
  ...
)
}
\arguments{
\item{gl}{A matrix of genotype log-likelihoods. The rows index the
individuals and the columns index the possible genotypes. So
\code{gl[i, k]} is the genotype log-likelihood for individual i and
genotype k-1.}

\item{p1_gl}{The vector of genotype likelihoods of parent 2.}

\item{mixprop}{The mixing proportion with the uniform for mixing purposes.}

\item{lg}{A logical. Should we log the marginal likelihood (\code{TRUE})
or not (\code{FALSE})?}

\item{...}{Additional parameters sent to \code{\link[stan]{sampling}()}.}
}
\description{
Marginal likelihood, double reduction, no preferential pairing, genotype likelihoods.
}
\examples{
\dontrun{
## null sims
set.seed(1)
gf <- offspring_gf(alpha = 0, xi = 1/3, p1 = 2, p2 = 2)
gcount <- offspring_geno(x = gf, n = 20)
gvec <- gcount_to_gvec(gcount)
fout <- po_gl(genovec = gvec, p1_geno = 2, p2_geno = 2, ploidy = 4)
gl <- fout$genologlike
mnull <- marg_f1_dr_npp_gl4(gl = gl, p1_gl = c(-Inf, -Inf, 0, -Inf, -Inf), p2_gl = c(-Inf, -Inf, 0, -Inf, -Inf), chains = 1)
malt <- -48.81 ## result of running marg_alt_gl(gl)
mnull - malt ## log-BF

## alt sims
set.seed(1)
gvec <- rep(0:4, each = 4)
fout <- po_gl(genovec = gvec, p1_geno = 2, p2_geno = 2, ploidy = 4)
gl <- fout$genologlike
mnull <- marg_f1_dr_npp_gl4(gl = gl, p1_gl = c(-Inf, -Inf, 0, -Inf, -Inf), p2_gl = c(-Inf, -Inf, 0, -Inf, -Inf), chains = 1)
malt <- -46.82 ## result of running marg_alt_gl(gl)
mnull - malt ## log-BF
}

}
\author{
Mira Thakkar and David Gerard
}
