% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glpknown.R
\name{marg_f1_dr_npp_glpknown4}
\alias{marg_f1_dr_npp_glpknown4}
\title{Marginal likelihood, double reduction, no preferential pairing, parent genotypes known, offspring genotypes unknown}
\usage{
marg_f1_dr_npp_glpknown4(gl, p1, p2, lg = TRUE, ...)
}
\arguments{
\item{gl}{A matrix of offspring genotype log-likelihoods. The rows index the
individuals and the columns index the possible genotypes. So
\code{gl[i, k]} is the offspring genotype log-likelihood for individual i and
genotype k-1.}

\item{p1}{Genotype of parent 1.}

\item{p2}{Genotype of parent 2.}

\item{lg}{A logical. Should we log the marginal likelihood (\code{TRUE}) or
not (\code{FALSE})?}

\item{...}{Additional paramters sent to \code{\link[stan]{sampling}()}.}
}
\description{
Marginal likelihood, double reduction, no preferential pairing, parent genotypes known, offspring genotypes unknown
}
\examples{
## null sims
set.seed(1)
gf <- offspring_gf(alpha = 0, xi = 1/3, p1 = 2, p2 = 2)
gcount <- offspring_geno(x = gf, n = 20)
gvec <- gcount_to_gvec(gcount)
fout <- po_gl(genovec = gvec, p1_geno = 2, p2_geno = 2, ploidy = 4)
gl <- fout$genologlike
mnull <- marg_f1_dr_npp_glpknown4(gl = gl, p1 = 2, p2 = 2, chains = 1)
malt <- -48.81 ## result of running marg_alt_gl(gl)
mnull - malt ## log-BF

## alt sims
set.seed(1)
gvec <- rep(0:4, each = 4)
fout <- po_gl(genovec = gvec, p1_geno = 2, p2_geno = 2, ploidy = 4)
gl <- fout$genologlike
mnull <- marg_f1_dr_npp_glpknown4(gl = gl, p1 = 2, p2 = 2, chains = 1)
malt <- -46.82 ## result of running marg_alt_gl(gl)
mnull - malt ## log-BF

}
\author{
Mira Thakkar and David Gerard
}
