% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glpknown.R
\name{marg_f1_dr_pp_glpknown4}
\alias{marg_f1_dr_pp_glpknown4}
\title{Marginal likelihood under null using offspring genotype log-likelihoods and parent genotypes}
\usage{
marg_f1_dr_pp_glpknown4(gl, p1, p2, lg = TRUE, ...)
}
\arguments{
\item{gl}{A matrix of offspring genotype log-likelihoods. The rows index the
individuals and the columns index the possible genotypes. So
\code{gl[i, k]} is the offspring genotype log-likelihood for individual i and
genotype k-1.}

\item{p1}{Genotype of parent 1.}

\item{p2}{Genotype of parent 2.}
}
\description{
Marginal likelihood under null using offspring genotype log-likelihoods and parent genotypes
}
\examples{
gl <- matrix(runif(20), ncol = 5)
marg_f1_dr_pp_gl4(gl = gl, p1 = 2, p2 = 3)

}
