% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_ploidy.R
\name{marg_f1_g}
\alias{marg_f1_g}
\title{Marginal likelihood for F1 populations of any ploidy.}
\usage{
marg_f1_g(x, beta = rep(1, (length(x) + 1)/2), mixprop = 0.001, lg = TRUE, ...)
}
\arguments{
\item{x}{The genotype counts of the offspring. \code{x[i]} is the
number of offspring with genotype \code{i-1}.}

\item{beta}{The concentration parameters for the dirichlet of the gamere
frequencies.}

\item{mixprop}{The mixing proportion with the uniform for mixing purposes.}

\item{lg}{A logical. Should we log the marginal likelihood (\code{TRUE})
or not (\code{FALSE})?}

\item{...}{Additional arguments to pass to \code{\link[rstan]{sampling}()}.}
}
\description{
Marginal likelihood for F1 populations of any ploidy.
}
\examples{
x <- c(4, 8, 12, 16, 12, 8, 4) ## generated under null
mnull <- marg_f1_g(x = x)
malt <- marg_alt_g(x = x)
mnull - malt

x <- rep(9, 7) ## generated under alt
mnull <- marg_f1_g(x = x)
malt <- marg_alt_g(x = x)
mnull - malt

}
\author{
David Gerard
}
