% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gknown.R
\name{marg_f1_g4}
\alias{marg_f1_g4}
\title{Marginal likelihood in F1 population of tetraploids}
\usage{
marg_f1_g4(
  x,
  p1_gl = rep(log(0.2), 5),
  p2_gl = rep(log(0.2), 5),
  lg = TRUE,
  ...
)
}
\arguments{
\item{x}{The genotype counts}

\item{p1_gl}{A vector of parent 1's genotype log-likelihoods.}

\item{p2_gl}{A vector of parent 2's genotype log-likelihoods}

\item{lg}{A logical. Should we log the marginal likelihood or not?}

\item{...}{Additional arguments to pass to \code{\link[rstan]{sampling}()}.}
}
\description{
Marginal likelihood in F1 population of tetraploids
}
\examples{
set.seed(1)
q <- hwep::zygdist(alpha = 0.1, G1 = 2, G2 = 2, ploidy = 4)
x <- c(stats::rmultinom(n = 1, size = 10, prob = q))
p1_gl <- rep(log(0.2), 5)
p2_gl <- rep(log(0.2), 5)

}
\author{
David Gerard
}
