% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gknown.R
\name{marg_f1_g4_pknown}
\alias{marg_f1_g4_pknown}
\title{Marginal likelihood in tetraploid F1 population}
\usage{
marg_f1_g4_pknown(x, g1, g2, lg = TRUE, ...)
}
\arguments{
\item{x}{The genotype counts}

\item{g1}{The first parent's genotype.}

\item{g2}{The second parent's gentoype.}

\item{lg}{A logical. Should we log the marginal likelihood or not?}

\item{...}{Additional arguments to pass to \code{\link[rstan]{sampling}()}.}
}
\description{
Here, parental genotypes are known.
}
\examples{
q <- hwep::zygdist(alpha = 0.1, G1 = 2, G2 = 2, ploidy = 4)
q <- runif(5)
q <- q / sum(q)
x <- c(stats::rmultinom(n = 1, size = 1000, prob = q))
g1 <- 2
g2 <- 2
marg_null <- marg_f1_g4_pknown(x = x, g1 = g1, g2 = g2, chains = 1)
marg_alt <- hwep:::ddirmult(x = x, alpha = rep(1, 5), lg = TRUE)
marg_null - marg_alt

}
\author{
David Gerard
}
