% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.R
\name{marg_f1_gl4}
\alias{marg_f1_gl4}
\title{Marginal likelihood under null using genotype log-likelihoods}
\usage{
marg_f1_gl4(
  gl,
  p1_gl = rep(log(0.2), 5),
  p2_gl = rep(log(0.2), 5),
  mixprop = 0.001,
  lg = TRUE,
  ...
)
}
\arguments{
\item{gl}{A matrix of genotype log-likelihoods. The rows index the
individuals and the columns index the possible genotypes. So
\code{gl[i, k]} is the genotype log-likelihood for individual i and
genotype k-1.}

\item{p1_gl}{A vector of parent 1's genotype log-likelihoods.}

\item{p2_gl}{A vector of parent 2's genotype log-likelihoods.}

\item{mixprop}{Mixing proportion with uniform to avoid
numerical issues in stan.}

\item{lg}{A logical. Should we log the marginal likelihood or not?}

\item{...}{Additional arguments to pass to \code{\link[rstan]{sampling}()}.}
}
\description{
Marginal likelihood under null using genotype log-likelihoods
}
\examples{
gl <- matrix(runif(20), ncol = 5)
marg_f1_gl4(gl = gl)

}
