% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gknown.R
\name{marg_f1_ndr_npp_g4}
\alias{marg_f1_ndr_npp_g4}
\title{Marginal likelihood, no double reduction, no preferential pairing, genotypes known.}
\usage{
marg_f1_ndr_npp_g4(x, g1, g2, lg = TRUE)
}
\arguments{
\item{x}{The genotype counts of the offspring. \code{x[i]} is the
number of offspring with genotype \code{i-1}.}

\item{g1}{The first parent's genotype.}

\item{g2}{The second parent's genotype.}

\item{lg}{A logical. Should we log the marginal likelihood (\code{TRUE})
or not (\code{FALSE})?}
}
\description{
Marginal likelihood, no double reduction, no preferential pairing, genotypes known.
}
\examples{
x <- c(3L, 21L, 52L, 20L, 4L) ## simulated via null
g1 <- 2
g2 <- 2
mnull <- marg_f1_ndr_npp_g4(x = x, g1 = g1, g2 = g2)
malt <- -15.34 ## output of marg_alt_g(x = x)
mnull - malt ## log-BF

x <- c(25L, 24L, 20L, 15L, 16L) ## simulated via alt
g1 <- 2
g2 <- 2
mnull <- marg_f1_ndr_npp_g4(x = x, g1 = g1, g2 = g2)
malt <- -15.34 ## output of marg_alt_g(x = x)
mnull - malt ## log-BF

}
\author{
Mira Thakkar and David Gerard
}
