% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gknown.R
\name{marg_f1_ndr_pp_g4}
\alias{marg_f1_ndr_pp_g4}
\title{Marginal likelihood, no double reduction, preferential pairing, genotypes known.}
\usage{
marg_f1_ndr_pp_g4(
  x,
  g1,
  g2,
  mixprop = 0.001,
  lg = TRUE,
  output = c("marg", "all"),
  ...
)
}
\arguments{
\item{x}{The genotype counts of the offspring. \code{x[i]} is the
number of offspring with genotype \code{i-1}.}

\item{g1}{The first parent's genotype.}

\item{g2}{The second parent's genotype.}

\item{mixprop}{The mixing proportion with the uniform for mixing purposes.}

\item{lg}{A logical. Should we log the marginal likelihood (\code{TRUE})
or not (\code{FALSE})?}

\item{output}{Return either the marginal likelihood with (\code{"marg"})
or the full STAN output with (\code{"all"}).}

\item{...}{Additional arguments to pass to \code{\link[rstan]{sampling}()}.}
}
\description{
Marginal likelihood, no double reduction, preferential pairing, genotypes known.
}
\examples{
x <- c(3L, 21L, 52L, 20L, 4L) ## simulated via null
g1 <- 2
g2 <- 2
mnull <- marg_f1_ndr_pp_g4(x = x, g1 = g1, g2 = g2)
malt <- -15.34 ## output of marg_alt_g(x = x)
mnull - malt ## log-BF

x <- c(25L, 24L, 20L, 15L, 16L) ## simulated via alt
g1 <- 2
g2 <- 2
mnull <- marg_f1_ndr_pp_g4(x = x, g1 = g1, g2 = g2)
malt <- -15.34 ## output of marg_alt_g(x = x)
mnull - malt ## log-BF

}
\author{
Mira Thakkar and David Gerard
}
