% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.R
\name{marg_f1_ndr_pp_gl4}
\alias{marg_f1_ndr_pp_gl4}
\title{Marginal likelihood, no double reduction, preferential pairing, genotype likelihoods.}
\usage{
marg_f1_ndr_pp_gl4(
  gl,
  p1_gl = rep(log(0.2), 5),
  p2_gl = rep(log(0.2), 5),
  mixprop = 0.001,
  lg = TRUE,
  ...
)
}
\arguments{
\item{gl}{A matrix of genotype log-likelihoods. The rows index the
individuals and the columns index the possible genotypes. So
\code{gl[i, k]} is the genotype log-likelihood for individual i and
genotype k-1.}

\item{p1_gl}{The vector of genotype likelihoods of parent 2.}

\item{mixprop}{The mixing proportion with the uniform for mixing purposes.}

\item{lg}{A logical. Should we log the marginal likelihood (\code{TRUE})
or not (\code{FALSE})?}

\item{...}{Additional paramters sent to \code{\link[stan]{sampling}()}.}
}
\description{
Marginal likelihood, no double reduction, preferential pairing, genotype likelihoods.
}
\examples{
gl <- matrix(runif(20), ncol = 5)
marg_f1_dr_npp_gl4(gl = gl)

}
\author{
Mira Thakkar and David Gerard
}
