% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menbayes-package.R
\docType{package}
\name{menbayes-package}
\alias{menbayes}
\alias{menbayes-package}
\title{menbayes: Bayesian Tests for Mendelian Segregation in Autopolyploids}
\description{
Provides a suite of methods for Bayesian tests for Mendelian segregation in experimental autopolyploid populations.
}
\author{
\strong{Maintainer}: David Gerard \email{gerard.1787@gmail.com} (\href{https://orcid.org/0000-0001-9450-5023}{ORCID})

Authors:
\itemize{
  \item Mira Thakkar \email{mt7562a@american.edu }
}

}
\keyword{internal}
