% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscripts.R
\name{offspring_gf}
\alias{offspring_gf}
\title{Function that takes as input the double reduction rate, the preferential
pairing rate, and parent genotypes to return zygote genotype frequencies.}
\usage{
offspring_gf(alpha, xi, p1, p2)
}
\arguments{
\item{alpha}{The double reduction rate}

\item{xi}{The preferential pairing parameter}

\item{p1}{The first parent's genotype}

\item{p2}{The second parent's genotype}
}
\value{
Zygote genotype frequencies
}
\description{
Function that takes as input the double reduction rate, the preferential
pairing rate, and parent genotypes to return zygote genotype frequencies.
}
\examples{
alpha <- 1/6
xi <- 1/3
p1 <- 2
p2 <- 3
offspring_gf(alpha = alpha, xi = xi, p1 = p1, p2 = p2)

}
\author{
Mira Thakkar
}
