% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscripts.R
\name{offspring_gf_2}
\alias{offspring_gf_2}
\title{Calculates zygote frequencies under the two-parameter model.}
\usage{
offspring_gf_2(alpha, xi, p1, p2)
}
\arguments{
\item{alpha}{The double reduction rate}

\item{xi}{The preferential pairing parameter}

\item{p1}{The first parent's genotype}

\item{p2}{The second parent's genotype}
}
\value{
Zygote genotype frequencies
}
\description{
Calculates zygote frequencies under the two-parameter model.
}
\examples{
alpha <- 1/6
xi <- 1/3
p1 <- 2
p2 <- 3
offspring_gf_2(alpha = alpha, xi = xi, p1 = p1, p2 = p2)

}
\author{
Mira Thakkar
}
