% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscripts.R
\name{offspring_gf_3}
\alias{offspring_gf_3}
\title{Calculates zygote frequencies under the two-parameter model.}
\usage{
offspring_gf_3(tau, beta, gamma, p1, p2)
}
\arguments{
\item{tau}{Probability of quadrivalent formation}

\item{beta}{Probability of double reduction given quadrivalent formation}

\item{gamma}{Probability of AA/aa pairing given bivalent formation}

\item{p1}{The first parent's genotype}

\item{p2}{The second parent's genotype}
}
\value{
Zygote genotype frequencies
}
\description{
Calculates zygote frequencies under the two-parameter model.
}
\examples{
offspring_gf_3(tau = 1/2, beta = 1/6, gamma = 1/3, p1 = 1, p2 = 2)

}
\author{
David Gerard
}
