% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{plot_beta}
\alias{plot_beta}
\title{Plot the beta distribution with 95\% area interval.}
\usage{
plot_beta(a, b, level = 0.8, ...)
}
\arguments{
\item{a}{the first shape parameter}

\item{b}{the second shape parameter}

\item{level}{The leve of the interval.}

\item{...}{Additional arguments to pass to \code{\link[graphics]{plot}}.}
}
\value{
Plots the beta density with \code{level} interval shaded, along
with the vertical line of the mean.
}
\description{
Plot the beta distribution with 95\% area interval.
}
\examples{
plot_beta(20, 40)
plot_beta(1/5, 2/5, ylim = c(0, 5))

}
\author{
David Gerard
}
