% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymapr.R
\name{polymapr_test}
\alias{polymapr_test}
\title{Run segregation distortion tests as implemented in the polymapR package.}
\usage{
polymapr_test(g, type = c("menbayes", "polymapR"))
}
\arguments{
\item{g}{Either a vector of genotype counts, or a matrix of genotype
likelihoods where the rows index the individuals and the columns
index the genotypes.}

\item{type}{Either my implementation which approximates that of
polymapR (\code{"menbayes"}) or the exact implementation
through polymapR (\code{"polymapR"}). Note that
polymapR needs to be installed for \code{type = "polymapR"}.}
}
\description{
The polymapR tests for segregation distortion by iterating through all
possible forms of disomic or polysomic inheritance from either parent,
tests for concordance of the offspring genotypes using a chi-squared
test, and returns the largest p-value. It sometimes chooses a different
p-value based on other heuristics. When \code{type = "menbayes"},
we only look at patterns of the given parent genotypes, choosing
the largest p-value. When \code{type = "polymapR"}, we return what
they use via their heuristics.
}
\author{
David Gerard
}
