% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscripts.R
\name{pvec}
\alias{pvec}
\title{Genotype frequencies of gametes when one parent's genotype is known}
\usage{
pvec(a, x, p)
}
\arguments{
\item{a}{The double reduction rate}

\item{x}{The preferential pairing parameter}

\item{p}{The}
}
\value{
The gamete genotype frequency
}
\description{
Genotype frequencies of gametes when one parent's genotype is known
}
\examples{
a <- 1/6
x <- 1/3
p <- 2
pvec(a = a, x = x, p = p)

}
\author{
Mira Thakkar
}
