% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscripts.R
\name{pvec_hex}
\alias{pvec_hex}
\title{Hexaploid genotype frequencies of gametes when one parent's genotype is known}
\usage{
pvec_hex(alpha, xi, ell)
}
\arguments{
\item{alpha}{The double reduction rate}

\item{xi}{The preferential pairing parameter}

\item{ell}{The parental genotype}
}
\description{
Hexaploid genotype frequencies of gametes when one parent's genotype is known
}
\examples{
alpha <- 3/10
xi <- 0.5
pvec_hex(alpha = alpha, xi = xi, ell = 0)
pvec_hex(alpha = alpha, xi = xi, ell = 1)
pvec_hex(alpha = alpha, xi = xi, ell = 2)
pvec_hex(alpha = alpha, xi = xi, ell = 3)
pvec_hex(alpha = alpha, xi = xi, ell = 4)
pvec_hex(alpha = alpha, xi = xi, ell = 5)
pvec_hex(alpha = alpha, xi = xi, ell = 6)


}
\author{
David Gerard
}
