% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gknown.R
\name{bayes_men_g4}
\alias{bayes_men_g4}
\title{Bayesian test for segregation distortion in tetraploids when genotypes are known.}
\usage{
bayes_men_g4(x, g1, g2, drbound = 1/6, pp = TRUE, dr = TRUE)
}
\arguments{
\item{x}{A vector of genotype counts. \code{x[i]} is the number of
offspring with genotype \code{i-1}.}

\item{g1}{The genotype of parent 1.}

\item{g2}{The genotype of parent 2.}

\item{drbound}{The maximum rate of double reduction. A default of 1/6
is provided, which is the rate under the complete equational
segregation model of meiosis.}

\item{pp}{A logical. Should we account for preferential pairing
(\code{TRUE}) or not (\code{FALSE})?}

\item{dr}{A logical. Should we account for double reduction
(\code{TRUE}) or not (\code{FALSE})?}
}
\description{
Bayesian test for segregation distortion in tetraploids when genotypes are known.
}
\examples{
set.seed(100)
gf <- offspring_gf_2(alpha = 1/12, xi1 = 0.2, xi2 = 0.6, p1 = 2, p2 = 2)
x <- offspring_geno(gf = gf, n = 100)
bayes_men_g4(x = x, g1 = 2, g2 = 2)

}
\author{
David Gerard
}
