% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glpknown.R
\name{bayes_men_gl4}
\alias{bayes_men_gl4}
\title{Bayesian test for segregation distortion in tetraploids using genotype likelihoods}
\usage{
bayes_men_gl4(
  gl,
  g1,
  g2,
  drbound = 1/6,
  shape1 = 5/9,
  shape2 = 10/9,
  ts1 = 1,
  ts2 = 1,
  pp = TRUE,
  dr = TRUE,
  xi = 1/3,
  alpha = 0,
  ...
)
}
\arguments{
\item{gl}{The genotype log-likelihoods. The rows index the individuals
and the columns index the genotypes.}

\item{g1}{Either parent 1's genotype, or parent 1's genotype log-likelihoods.}

\item{g2}{Either parent 2's genotype, or parent 2's genotype log-likelihoods.}

\item{drbound}{The maximum rate of double reduction. A default of 1/6
is provided, which is the rate under the complete equational
segregation model of meiosis.}

\item{shape1}{The shape 1 parameter for the beta prior over the
preferential pairing parameter.}

\item{shape2}{The shape 2 parameter for the beta prior over the
preferential pairing parameter.}

\item{ts1}{The shape 1 parameter for the beta prior over the
rate of quadrivalent formation.}

\item{ts2}{The shape 2 parameter for the beta prior over the
rate of quadrivalent formation.}

\item{pp}{A logical. Should we account for preferential pairing
(\code{TRUE}) or not (\code{FALSE})?}

\item{dr}{A logical. Should we account for double reduction
(\code{TRUE}) or not (\code{FALSE})?}

\item{xi}{If \code{pp = FALSE}, then \code{xi} is the known rate of
preferential pairing (assumed fixed).}

\item{alpha}{If \code{dr = FALSE}, then \code{alpha} is the known rate of
double reduction (assumed fixed).}

\item{...}{Additional arguments to pass to Stan.}
}
\description{
Bayesian test for segregation distortion in tetraploids using genotype likelihoods
}
\section{Unidentified parameters}{

When \code{g1 = 2} or \code{g2 = 2} (or both), the model is not identified
and those estimates (\code{alpha}, \code{xi1}, and \code{xi2}) are
meaningless. Do NOT interpret them. The log-BF is fine, though.
}

\examples{
\dontrun{
## null sims
set.seed(1)
g1 <- 1
g2 <- 2
gl <- simf1gl(n = 25, g1 = g1, g2 = g2)
bayes_men_gl4(gl = gl, g1 = g1, g2 = g2, chains = 1, iter = 1000)

## genotypes of parents not known
g1_gl <- log(c(0.2, 0.5, 0.2, 0.09, 0.01))
g2_gl <- log(c(0.09, 0.2, 0.4, 0.3, 0.01))
bayes_men_gl4(gl = gl, g1 = g1_gl, g2 = g2_gl, chains = 1, iter = 1000)

## alt sims
gl <- hwep::simgl(nvec = rep(5, 5), rdepth = 10)
bayes_men_gl4(gl = gl, g1 = g1, g2 = g2, chains = 1, iter = 1000)

bayes_men_gl4(gl = gl, g1 = g1_gl, g2 = g2_gl, chains = 1, iter = 1000)
}

}
\author{
David Gerard
}
