% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gknown.R
\name{marg_alt_g}
\alias{marg_alt_g}
\title{Marginal likelihood under alternative when genotypes are known}
\usage{
marg_alt_g(x, beta = rep(1, length(x)), lg = TRUE, ...)
}
\arguments{
\item{x}{The genotype counts}

\item{beta}{The prior hyperparameters}

\item{lg}{A logical. Should we log the marginal likelihood or not?}

\item{...}{Additional arguments to pass to \code{\link[rstan]{sampling}()}.}
}
\value{
The marginal likelihood.
}
\description{
Marginal likelihood under alternative when genotypes are known
}
\examples{
x <- c(1, 2, 7, 3, 3)
beta <- rep(1, 5)
marg_alt_g(x = x, beta = beta, chains = 1)

}
\author{
David Gerard
}
