% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscripts.R
\name{offspring_geno}
\alias{offspring_geno}
\title{A function called offspring_geno() Which takes as input the offspring
genotype frequencies and a sample size and returns simulated genotypes.}
\usage{
offspring_geno(x, n)
}
\arguments{
\item{x}{Vector of offspring genotype frequencies}

\item{n}{Sample size}
}
\value{
Simulated genotypes
}
\description{
A function called offspring_geno() Which takes as input the offspring
genotype frequencies and a sample size and returns simulated genotypes.
}
\examples{
x <- offspring_gf(alpha = 1/6, xi = 1/3, p1 = 2, p2 = 3)
n <- 1000
pvec(x = x, n = n)

}
\author{
Mira Thakkar
}
